S.S. Anne healer and other bits
This adds the S.S. Anne healer from FRLG that's notorious for gaslighting newer players who just think she's there.

This also adds an extremely early Mewtwo theme. It's far from finished, just happens to be in this commit as I was testing it in debug - thanks to Luci for working on this!

Updated bug list as well with things my friend Crunter found.