Fix Mewtwo theme play issues
Mewtwo's theme wasn't playing properly because I am silly and forgot to add a `jr .playSong` so it doesn't fall through. Also, it needed to be before the wild check. Now it works as expected.

This also has the latest version of the Mewtwo theme by LuciShrimp, which is reaching a pretty amazing state.