Finish Post-Game Letter & Citrine Pass
This makes it so that once you clear the game, you will be able to get the Silph Letter from your mom, and the Citrine Pass by beating Professor Oak.

You'll now walk to your mother upon leaving your room post-clearing the game, and she'll hand you the letter. This'll basically explain your next quest to you.

The Professor Oak event kind of destroyed my sanity but that's fine. It works. Thanks to Martha for the idea.

Considerations for a full bag are done and they were annoying. I decided to show off and do some comedy to alleviate the lost sanity.