Fix crash when losing to Yujirou
The reason was because not all the infrastructure for an actual battle was set up - this fixes it. It still has some strange text that appears though.

Also did some minor optimisations on the Vermillion Dock stuff - I think the direction I am taking Faraway Island will work. However, something in that script causes any and all special warps to crash in some way. I am not sure what it is that does it.