Populating Brunswick Trail
This adds three trainers to Brunswick Trail, as well as the lover NPCs Martha wanted to write dialogue for (with placeholders for now).

Most trainers are merely commenting on the strong and energetic Pokemon, which I think matches the racing theme it has going on. This can be changed if need-be.

I have also removed all the noise on the Fake Tree sprite that got added when I imported the tile. Now, it's far more convincing.