Wild data refinements
Some of the wild data on these routes needs to be looked at long and hard. Many of them have an extreme amount of variance that ends up making it harder for players to find what they actually want - note that in these games, it does take a while to get through a single one. In general, reduce the variance. Some areas feel a bit chaotic as well, I noticed Barunda in a few places that felt off.

Victory Road has been refined, with the Hitmon repeats made more accurate to LGPE - they are encountered individually on different floors. I have upgraded much of the area's residents, such as adding Rhydon from FRLG and turning Onix, Graveler, and Marowak into Trampel, Crocky, and Guardia. Steelix is a worthy consideration. I've kept Machoke and Golbat to serve as "blanks" to KO for exp on weaker Pokemon the player may have. It could still be refined, but this is a step in the right direction.

I've decided to put Kotora in the "Pikachu slot" in Viridian Forest as a compromise for the lack of an early Electric-type when Pikachu isn't chosen. It's at Level 3, so it will take longer to train by comparison - thus, by this point, it cannot compete with the starter Pikachu, as that will be around Level 10-12.