Legendary Theme Demake
It's done! This took a lot of work on LuciShrimp's end, mainly due to CrystalTracker issues.

This fixes the synchronisation issues seen in the previous iteration, dramatically improves the loop point, and has various other shifts to make it much easier on the ears.

I'm glad we chose to use the Mew version of the theme from Emerald, it feels much more complete!