Finalise Mt. Moon Square placement
This puts Mt. Moon Square into an area solidly reminiscent of GSC. To prevent breaking of progression, a Cut tree is placed where the exit out of Mt. Moon is.

Also, I replaced a Potion with a visible Moon Stone - felt awkward not having one. It's a bit out of the way, so players have to explore, which feels correct.