Hall of Fame Crash Fix
I forgot to put a `ret` at the end of `ShowThis`, causing the sequence to break. It also fixes a bug that would trap you in your house if you cleared the game at all - LAST_MAP is set to the player's room due to how we changed the starting position.

The post-game Legendary respawns are NOT fixed with this, for reasons I don't yet know. It's most likely the faulty dex as a whole.

The post-game setup itself now works, though - Route 1 Oak appears, Crater Guard disappears, Bill's Garden becomes accessible, etc.

Cerulean Cave is made accessible through the post-game setup until the gauntlet is officially done.