Re-frame Red/Green character options
This re-frames the intro to have Red and Green be framed in a gender neutral manner. The main aim here is to be a bit more appealing to the NB types, be more consistent with the modern game's approach to gender, and allow people to feel comfier with Red's androgynous appearance.

Oh, and there's a confirmation option to utilise this more effectively, along with some cool transitions. Very happy with this commit.