3DS idiot-proofing & other bits
This adds a makeshift border, which will later be replaced with a proper KEP one (thanks Gecko557) The readme also includes a link to the software itself.

Responding to some feedback on YouTube, the Citrine Mart now features PP Ups and Max Elixers as renewable items. PP Ups have never been renewable!

Oh, and the Exp. All is now a key item because for some reason it wasn't before. Yikes.

Last commit for the night.